requirements.txt
openai
tqdm
python-dotenv
fastapi
uvicorn
httpx
pydantic